These got missed on the previous commit.
Adding some new details to README.md and some links to tools
we may use.

Putting grunt/plugins into dev-dependencies section of
package.json

Added "Super Strict" .jshintrc as a starting point, which we
will relax.

Gruntfile.js is a curious format... wonder if Tom wants
this or was using and existing one?