Going back to basics... should have done on a different branch.
Following along on the Pluralsight course for Grunt and build workflows
by Dave Mosher.

Stripped Gruntfile.js down but also scanning existing ones on disk to
check for standard (well, common) formatting, where they differ from
Daves.

Had to add the has-color module for some reason.

ignoring build subdir where we, for example, add the results of
concatenation, which is the first grunt task we are working with.

If I add a std local tasks dir and have it empty and call grunt
with a default to 'concat', the system bails because it looks
for a concat.js file in the locals 'tasks' dir.

Hopefully that will get resolved.

Also starting to add paths to sources etc that match the ones Tom
and I discussed.