We use bluebird to create a Promise for the resetJobs call.
resolve and reject will be called for pass/fail, so we can ignore our
callback.

Then when we call our function, we add a .then call (if the promise
was resolved (pass!)

This does not look much right now, but in a few commits it
will look a lot cleaner and will allowed chained promises.